requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
deep-translator==1.11.4
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-translate==3.16.0
googleapis-common-protos==1.65.0
googletrans==3.1.0a0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
lxml==5.3.0
oauthlib==3.2.2
outcome==1.3.0.post0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyparsing==3.2.0
PySocks==1.7.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
